LayoutTests/platform/ios-simulator/css3/device-adapt/opera/constrain-011-expected.txt


FAIL CSS Test: @viewport constrained - width from zoom: 200% assert_equals: expected 400 but got 800(stack: assert@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:2168:73
assert_equals@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:705:15
file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/constrain-011.xhtml:69:22
step@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/resources/testharness.js:1186:30
onload@file:///Volumes/Data/Development/OSX/webkit/OpenSource/LayoutTests/css3/device-adapt/opera/constrain-011.xhtml:68:14)
